+++ application/tests/models/Sites_Model_test.php
@@ -98,6 +98,13 @@ class Site_Model_test extends TestCase {
 		$this->assertRegExp('/^[0-9]+$/', $result['latest_chapter']);
 		$this->assertRegExp('/^[0-9]+-[0-9]+-[0-9]+ [0-9]+:[0-9]+:[0-9]+$/', $result['last_updated']);
 	}
+	public function test_MangaPanda_fail() {
+		MonkeyPatch::patchFunction('log_message', NULL, 'MangaPanda'); //Stop logging stuff...
+		$result = $this->Sites_Model->{'MangaPanda'}->getTitleData('i_am_a_bad_url');
+
+		$this->assertNull($result);
+		MonkeyPatch::verifyInvokedOnce('log_message', ['error', 'MangaPanda : i_am_a_bad_url | Bad Status Code (404)']);
+	}
 
 	public function test_MangaStream() {
 		$this->skipTravis('Travis\'s PHP Curl ver. doesn\'t seem to play nice with SSL.');